requirements.txt
streamlit>=1.37.0
langchain>=0.3.0
langgraph>=0.2.15
sentence-transformers>=3.0.1
transformers>=4.43.0
huggingface-hub>=0.23.0
torch>=2.2.0
accelerate>=1.0.0
duckduckgo-search>=6.1.0
wikipedia>=1.4.0
requests>=2.31.0
beautifulsoup4>=4.12.3
pdfplumber>=0.11.0
chromadb>=0.5.3
faiss-cpu>=1.8.0
playwright>=1.47.0
numpy>=1.26.4
scikit-learn>=1.4.2
tiktoken>=0.7.0
pytest>=8.2.0
pytest-mock>=3.14.0
